% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterData.R
\name{FilterData}
\alias{FilterData}
\title{Filter data using variance, coefficient of variation, or sparsity.}
\usage{
FilterData(
  xdata,
  ydata,
  col_var_percentile_filter = 0,
  row_var_percentile_filter = 0,
  col_coeffvar_percentile_filter = 0,
  row_coeffvar_percentile_filter = 0,
  col_sparsity_min_nonzero_percentile = 0,
  row_sparsity_min_nonzero_percentile = 0,
  remove_zero_median_cols = F,
  remove_zero_median_rows = F,
  scale_zeroes_directly = 0,
  remove_zero_sd_cols = T
)
}
\arguments{
\item{xdata}{the data matrix}

\item{ydata}{feature labels/values}

\item{col_var_percentile_filter}{double to use as lower percentile to use for filtering cols by variance. Uses percentile from
histogram breaks.}

\item{row_var_percentile_filter}{double to use as lower percentile to use for filtering rows by variance. Uses percentile from
histogram breaks.}

\item{col_coeffvar_percentile_filter}{double to use as lower percentile to use for filtering cols by coefficient of variation. Uses percentile from
histogram breaks.}

\item{row_coeffvar_percentile_filter}{double to use as lower percentile to use for filtering rows by coefficient of variation. Uses percentile from
histogram breaks.}

\item{col_sparsity_min_nonzero_percentile}{double to use as lower percentile to use for filtering columns by sparsity. Uses percentile from
histogram breaks.}

\item{row_sparsity_min_nonzero_percentile}{double to use as lower percentile to use for filtering rows by sparsity. Uses percentile from
histogram breaks.}

\item{remove_zero_median_cols}{logical for whether to remove cols with median = 0}

\item{remove_zero_median_rows}{logical for whether to remove rows with median = 0}

\item{scale_zeroes_directly}{double to use for replacing zeros in data matrix}

\item{remove_zero_sd_cols}{logical for whether to filter features with zero std dev.}
}
\description{
Filter data using variance, coefficient of variation, or sparsity.
}
