% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSigGenes.R
\name{plotSigGenes}
\alias{plotSigGenes}
\title{Plot significant genes for significant (marginal) latent factors}
\usage{
plotSigGenes(slide_results, out_path = NULL, annotate_anchors = F)
}
\arguments{
\item{slide_results}{list - SLIDE results output from runSLIDE function}

\item{out_path}{string - Output path to save results}

\item{annotate_anchors}{logical - whether to annotate the latent factor
with a '*'}
}
\value{
plt - ggplot image

plot_df - dataframe used to plot results
}
\description{
Plot significant genes for significant (marginal) latent factors
}
